crashpod.v2/code/requirements.txt
Flask==1.1.4
Werkzeug
markupsafe==2.0.1
prometheus-flask-instrumentator
requests
python-dateutil
cryptography==39.0.1